+++ lib/src/base_dropdown_search.dart
@@ -18,17 +18,24 @@ import 'package:flutter/services.dart';
 
 import 'widgets/custom_scroll_view.dart';
 
-typedef DropdownSearchOnFind<T> = FutureOr<List<T>> Function(String filter, LoadProps? loadProps);
+typedef DropdownSearchOnFind<T> = FutureOr<List<T>> Function(
+    String filter, LoadProps? loadProps);
 typedef DropdownSearchItemAsString<T> = String Function(T item);
 typedef DropdownSearchFilterFn<T> = bool Function(T item, String filter);
 typedef DropdownSearchCompareFn<T> = bool Function(T item1, T item2);
-typedef DropdownSearchBuilder<T> = Widget Function(BuildContext context, T? selectedItem);
-typedef DropdownSearchBuilderMultiSelection<T> = Widget Function(BuildContext context, List<T> selectedItems);
-typedef DropdownSearchPopupItemBuilder<T> = Widget Function(BuildContext context, T item, bool isDisabled, bool isSelected);
+typedef DropdownSearchBuilder<T> = Widget Function(
+    BuildContext context, T? selectedItem);
+typedef DropdownSearchBuilderMultiSelection<T> = Widget Function(
+    BuildContext context, List<T> selectedItems);
+typedef DropdownSearchPopupItemBuilder<T> = Widget Function(
+    BuildContext context, T item, bool isDisabled, bool isSelected);
 typedef DropdownSearchPopupItemEnabled<T> = bool Function(T item);
-typedef ErrorBuilder = Widget Function(BuildContext context, String searchEntry, dynamic exception);
-typedef EmptyBuilder = Widget Function(BuildContext context, String searchEntry);
-typedef LoadingBuilder = Widget Function(BuildContext context, String searchEntry);
+typedef ErrorBuilder = Widget Function(
+    BuildContext context, String searchEntry, dynamic exception);
+typedef EmptyBuilder = Widget Function(
+    BuildContext context, String searchEntry);
+typedef LoadingBuilder = Widget Function(
+    BuildContext context, String searchEntry);
 typedef BeforeChange<T> = FutureOr<bool> Function(T? prevItem, T? nextItem);
 typedef BeforePopupOpening<T> = FutureOr<bool> Function(T? selectedItem);
 typedef BeforePopupOpeningMultiSelection<T> = FutureOr<bool> Function(List<T> selectedItem);
@@ -40,9 +47,11 @@ typedef OnItemAdded<T> = void Function(List<T> selectedItems, T addedItem);
 typedef OnItemRemoved<T> = void Function(List<T> selectedItems, T removedItem);
 typedef OnClearMultiSelection<T> = void Function();
 typedef OnClear<T> = void Function();
-typedef OnBeforeClearMultiSelection<T> = FutureOr<bool> Function(List<T> selectedItems);
+typedef OnBeforeClearMultiSelection<T> = FutureOr<bool> Function(
+    List<T> selectedItems);
 typedef OnBeforeClear<T> = FutureOr<bool> Function(T? selectedItem);
-typedef ContainerBuilder<T> = Widget Function(BuildContext context, Widget child);
+typedef ContainerBuilder<T> = Widget Function(
+    BuildContext context, Widget child);
 
 enum PopupMode { dialog, modalBottomSheet, menu, bottomSheet, autocomplete }
 
@@ -202,7 +211,8 @@ abstract class BaseDropdownSearch<T> extends StatefulWidget {
           'Please implement your `dropdownBuilder`',
         ),
         assert(
-          mode != Mode.custom || (decoratorProps == null && onSaved == null && validator == null),
+          mode != Mode.custom ||
+              (decoratorProps == null && onSaved == null && validator == null),
           'Custom mode has no form properties',
         ),
         assert(
@@ -268,7 +278,8 @@ abstract class BaseDropdownSearch<T> extends StatefulWidget {
           'Please implement your `dropdownBuilder`',
         ),
         assert(
-          mode != Mode.custom || (decoratorProps == null && onSaved == null && validator == null),
+          mode != Mode.custom ||
+              (decoratorProps == null && onSaved == null && validator == null),
           "Custom mode has no form properties",
         ),
         assert(
@@ -321,7 +332,8 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
     _selectedItemsNotifier.value = List.from(widget.selectedItems);
 
     if (widget.popupProps.mode == PopupMode.autocomplete) {
-      HardwareKeyboard.instance.addHandler(_handleAutoCompleteBackPressKeyPress);
+      HardwareKeyboard.instance
+          .addHandler(_handleAutoCompleteBackPressKeyPress);
     }
 
     _isFocused.addListener(() {
@@ -330,8 +342,10 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
   }
 
   bool _handleAutoCompleteBackPressKeyPress(KeyEvent event) {
-    if (event is KeyDownEvent && event.logicalKey == LogicalKeyboardKey.backspace) {
-      if (_popupStateKey.currentState?.searchBoxController.text.isEmpty == true) {
+    if (event is KeyDownEvent &&
+        event.logicalKey == LogicalKeyboardKey.backspace) {
+      if (_popupStateKey.currentState?.searchBoxController.text.isEmpty ==
+          true) {
         final item = getSelectedItems.last;
         removeItem(item);
         _popupStateKey.currentState?.deselectItems([item]);
@@ -349,7 +363,8 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
 
     ///this code check if we need to refresh the popup widget to update
     ///containerBuilder widget
-    if (widget.popupProps.containerBuilder != oldWidget.popupProps.containerBuilder) {
+    if (widget.popupProps.containerBuilder !=
+        oldWidget.popupProps.containerBuilder) {
       WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
         _popupStateKey.currentState?.setState(() {});
       });
@@ -376,7 +391,10 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
                     hoverColor: Colors.transparent,
                   )
                 : widget.clickProps,
-            onTap: () => widget.popupProps.mode == PopupMode.autocomplete && isFocused ? null : openDropDownSearch(),
+            onTap: () =>
+                widget.popupProps.mode == PopupMode.autocomplete && isFocused
+                    ? null
+                    : openDropDownSearch(),
             child: _dropDown(),
           ),
         );
@@ -389,7 +407,8 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
     closeDropDownSearch();
     autoCompleteFocusNode.dispose();
     _isFocused.dispose();
-    HardwareKeyboard.instance.removeHandler(_handleAutoCompleteBackPressKeyPress);
+    HardwareKeyboard.instance
+        .removeHandler(_handleAutoCompleteBackPressKeyPress);
     super.dispose();
   }
 
@@ -437,13 +456,17 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
                   child: isFocused
                       ? Row(
                           children: [
-                            if (selectedItems != null) Flexible(child: selectedItems),
-                            if (selectedItems != null) Padding(padding: EdgeInsets.only(left: 4)),
+                            if (selectedItems != null)
+                              Flexible(child: selectedItems),
+                            if (selectedItems != null)
+                              Padding(padding: EdgeInsets.only(left: 4)),
                             Expanded(
                               child: TextFormField(
                                 focusNode: autoCompleteFocusNode,
-                                controller: _popupStateKey.currentState?.searchBoxController,
-                                decoration: InputDecoration(border: InputBorder.none),
+                                controller: _popupStateKey
+                                    .currentState?.searchBoxController,
+                                decoration:
+                                    InputDecoration(border: InputBorder.none),
                               ),
                             ),
                           ],
@@ -476,11 +499,14 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
               props: ChipProps(
                 onDeleted: widget.enabled ? () => removeItem(e) : null,
                 shape: widget.uiMode == UiToApply.cupertino
-                    ? RoundedRectangleBorder(borderRadius: BorderRadius.all(Radius.circular(18)))
+                    ? RoundedRectangleBorder(
+                        borderRadius: BorderRadius.all(Radius.circular(18)))
                     : null,
                 padding: const EdgeInsets.all(0),
                 selected: true,
-                deleteIcon: widget.uiMode == UiToApply.cupertino ? Icon(CupertinoIcons.multiply_circle_fill) : null,
+                deleteIcon: widget.uiMode == UiToApply.cupertino
+                    ? Icon(CupertinoIcons.multiply_circle_fill)
+                    : null,
               ).merge(widget.chipProps),
             );
           }).toList(),
@@ -500,7 +526,8 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
     if (widget.dropdownBuilder != null) {
       selectedItem = widget.dropdownBuilder!(context, getSelectedItem);
     } else if (widget.dropdownBuilderMultiSelection != null) {
-      selectedItem = widget.dropdownBuilderMultiSelection!(context, getSelectedItems);
+      selectedItem =
+          widget.dropdownBuilderMultiSelection!(context, getSelectedItems);
     } else if (isMultiSelectionMode) {
       selectedItem = defaultSelectedItems(getSelectedItems);
     }
@@ -523,17 +550,21 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
   TextStyle? _getDecoratorBaseTextStyle() {
     return widget.enabled
         ? widget.decoratorProps.baseStyle
-        : TextStyle(color: Theme.of(context).disabledColor).merge(widget.decoratorProps.baseStyle);
+        : TextStyle(color: Theme.of(context).disabledColor)
+            .merge(widget.decoratorProps.baseStyle);
   }
 
   TextStyle? _getTextStyle() {
     return widget.enabled
         ? widget.textProps.style
-        : TextStyle(color: Theme.of(context).disabledColor).merge(widget.textProps.style);
+        : TextStyle(color: Theme.of(context).disabledColor)
+            .merge(widget.textProps.style);
   }
 
   Widget _formField() {
-    return isMultiSelectionMode ? _formFieldMultiSelection() : _formFieldSingleSelection();
+    return isMultiSelectionMode
+        ? _formFieldMultiSelection()
+        : _formFieldSingleSelection();
   }
 
   Widget _customField() => _buildSelectedItemsWidget() ?? SizedBox.shrink();
@@ -634,8 +665,12 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
     );
 
     final decoration = widget.decoratorProps.decoration ??
-        (widget.uiMode == UiToApply.cupertino ? cupertinoDecoration : materialDefaultDecoration);
-    return decoration.applyDefaults(Theme.of(state.context).inputDecorationTheme).copyWith(
+        (widget.uiMode == UiToApply.cupertino
+            ? cupertinoDecoration
+            : materialDefaultDecoration);
+    return decoration
+        .applyDefaults(Theme.of(state.context).inputDecorationTheme)
+        .copyWith(
           enabled: widget.enabled,
           suffixIcon: _manageSuffixIcons(),
           errorText: state.errorText,
@@ -663,17 +698,23 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
       return CustomIconButton(
         props: props,
         onPressed: () => clear(),
-        icon: props.icon ?? Icon(widget.uiMode == UiToApply.cupertino ? CupertinoIcons.clear_circled_solid : Icons.clear),
+        icon: props.icon ??
+            Icon(widget.uiMode == UiToApply.cupertino
+                ? CupertinoIcons.clear_circled_solid
+                : Icons.clear),
       );
     }
 
     Widget? getDropdownButton() {
-      final dropDownButton = widget.suffixProps.dropdownButtonProps ?? DropdownButtonProps();
+      final dropDownButton =
+          widget.suffixProps.dropdownButtonProps ?? DropdownButtonProps();
       if (!dropDownButton.isVisible) return null;
 
       //icon required
       final dropDownClosedIcon = dropDownButton.icon ??
-          Icon(widget.uiMode == UiToApply.cupertino ? CupertinoIcons.chevron_down : Icons.arrow_drop_down);
+          Icon(widget.uiMode == UiToApply.cupertino
+              ? CupertinoIcons.chevron_down
+              : Icons.arrow_drop_down);
 
       final icon = dropDownButton.iconOpened == null
           ? dropDownClosedIcon
@@ -730,15 +771,18 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
           constraints: widget.popupProps.constraints,
           onTapOutside: (event) => closeDropDownSearch(),
           props: isMultiSelectionMode
-              ? (widget.popupProps as CupertinoMultiSelectionPopupProps<T>).autoCompleteProps
-              : (widget.popupProps as CupertinoPopupProps<T>).autoCompleteProps);
+              ? (widget.popupProps as CupertinoMultiSelectionPopupProps<T>)
+                  .autoCompleteProps
+              : (widget.popupProps as CupertinoPopupProps<T>)
+                  .autoCompleteProps);
     } else {
       _customOverlyEntry = MaterialCustomOverlyEntry(
           child: _popupWidgetInstance(),
           constraints: widget.popupProps.constraints,
           onTapOutside: (event) => closeDropDownSearch(),
           props: isMultiSelectionMode
-              ? (widget.popupProps as MultiSelectionPopupProps<T>).autoCompleteProps
+              ? (widget.popupProps as MultiSelectionPopupProps<T>)
+                  .autoCompleteProps
               : (widget.popupProps as PopupProps<T>).autoCompleteProps);
     }
     return _customOverlyEntry!.open(context);
@@ -751,7 +795,8 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
         context,
         _popupWidgetInstance(),
         isMultiSelectionMode
-            ? (widget.popupProps as CupertinoMultiSelectionPopupProps<T>).dialogProps
+            ? (widget.popupProps as CupertinoMultiSelectionPopupProps<T>)
+                .dialogProps
             : (widget.popupProps as CupertinoPopupProps<T>).dialogProps,
         [
           CupertinoDialogAction(
@@ -784,7 +829,8 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
         context,
         _popupWidgetInstance(),
         isMultiSelectionMode
-            ? (widget.popupProps as CupertinoMultiSelectionPopupProps<T>).bottomSheetProps
+            ? (widget.popupProps as CupertinoMultiSelectionPopupProps<T>)
+                .bottomSheetProps
             : (widget.popupProps as CupertinoPopupProps<T>).bottomSheetProps,
       );
     }
@@ -805,9 +851,16 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
         context,
         _popupWidgetInstance(),
         isMultiSelectionMode
-            ? (widget.popupProps as CupertinoMultiSelectionPopupProps<T>).modalBottomSheetProps
-            : (widget.popupProps as CupertinoPopupProps<T>).modalBottomSheetProps,
-        isMultiSelectionMode ? [CupertinoActionSheetAction(onPressed: () => popupOnValidate(), child: Text('OK'))] : null,
+            ? (widget.popupProps as CupertinoMultiSelectionPopupProps<T>)
+                .modalBottomSheetProps
+            : (widget.popupProps as CupertinoPopupProps<T>)
+                .modalBottomSheetProps,
+        isMultiSelectionMode
+            ? [
+                CupertinoActionSheetAction(
+                    onPressed: () => popupOnValidate(), child: Text('OK'))
+              ]
+            : null,
       );
     }
 
@@ -815,7 +868,8 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
       context,
       _popupWidgetInstance(),
       isMultiSelectionMode
-          ? (widget.popupProps as MultiSelectionPopupProps<T>).modalBottomSheetProps
+          ? (widget.popupProps as MultiSelectionPopupProps<T>)
+              .modalBottomSheetProps
           : (widget.popupProps as PopupProps<T>).modalBottomSheetProps,
     );
   }
@@ -824,7 +878,8 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
   Future _openMenu() {
     if (widget.uiMode == UiToApply.cupertino) {
       final menuProps = isMultiSelectionMode
-          ? (widget.popupProps as CupertinoMultiSelectionPopupProps<T>).menuProps
+          ? (widget.popupProps as CupertinoMultiSelectionPopupProps<T>)
+              .menuProps
           : (widget.popupProps as CupertinoPopupProps<T>).menuProps;
 
       return CupertinoCustomMenu<T>(
@@ -888,7 +943,8 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
 
     if (widget.onBeforeChange != null) {
       Future.value(
-        widget.onBeforeChange!(getSelectedItem, selectedItems.isEmpty ? null : selectedItems.first),
+        widget.onBeforeChange!(getSelectedItem,
+            selectedItems.isEmpty ? null : selectedItems.first),
       ).then((allowed) => allowed ? changeItem() : null);
     } else if (widget.onBeforeChangeMultiSelection != null) {
       Future.value(
@@ -921,7 +977,8 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
     if (widget.onBeforePopupOpening != null) {
       if (await widget.onBeforePopupOpening!(getSelectedItem) == false) return;
     } else if (widget.onBeforePopupOpeningMultiSelection != null) {
-      if (await widget.onBeforePopupOpeningMultiSelection!(getSelectedItems) == false) {
+      if (await widget.onBeforePopupOpeningMultiSelection!(getSelectedItems) ==
+          false) {
         return;
       }
     }
@@ -942,17 +999,19 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
   ///Change selected Value; this function is public USED to change the selected
   ///value PROGRAMMATICALLY, Otherwise you can use [_handleOnChangeSelectedItems]
   ///for multiSelection mode you can use [changeSelectedItems]
-  void changeSelectedItem(T? selectedItem) => _handleOnChangeSelectedItems(BaseDropdownSearch._itemToList(selectedItem));
+  void changeSelectedItem(T? selectedItem) => _handleOnChangeSelectedItems(
+      BaseDropdownSearch._itemToList(selectedItem));
 
   ///Change selected Value; this function is public USED to change the selected
   ///value PROGRAMMATICALLY, Otherwise you can use [_handleOnChangeSelectedItems]
   ///for SingleSelection mode you can use [changeSelectedItem]
-  void changeSelectedItems(List<T> selectedItems) => _handleOnChangeSelectedItems(selectedItems);
+  void changeSelectedItems(List<T> selectedItems) =>
+      _handleOnChangeSelectedItems(selectedItems);
 
   ///function to remove an item from the list
   ///Useful in multiSelection mode to delete an item
-  void removeItem(T itemToRemove) =>
-      _handleOnChangeSelectedItems(getSelectedItems..removeWhere((i) => _isEqual(itemToRemove, i)));
+  void removeItem(T itemToRemove) => _handleOnChangeSelectedItems(
+      getSelectedItems..removeWhere((i) => _isEqual(itemToRemove, i)));
 
   ///Change selected Value; this function is public USED to clear selected
   ///value PROGRAMMATICALLY, Otherwise you can use [_handleOnChangeSelectedItems]
@@ -976,7 +1035,8 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
   }
 
   ///get selected value programmatically USED for SINGLE_SELECTION mode
-  T? get getSelectedItem => getSelectedItems.isEmpty ? null : getSelectedItems.first;
+  T? get getSelectedItem =>
+      getSelectedItems.isEmpty ? null : getSelectedItems.first;
 
   ///get selected values programmatically
   List<T> get getSelectedItems => _selectedItemsNotifier.value;
@@ -1046,13 +1106,16 @@ class DropdownSearchState<T> extends State<BaseDropdownSearch<T>> {
   }
 
   ///returns true if all popup's items are selected; other wise False
-  bool get popupIsAllItemSelected => _popupStateKey.currentState?.isAllItemSelected ?? false;
+  bool get popupIsAllItemSelected =>
+      _popupStateKey.currentState?.isAllItemSelected ?? false;
 
   ///returns popup selected items
-  List<T> get popupGetSelectedItems => _popupStateKey.currentState?.getSelectedItem ?? [];
+  List<T> get popupGetSelectedItems =>
+      _popupStateKey.currentState?.getSelectedItem ?? [];
 
   ///returns popup showed/loaded items
-  List<T> get popupGetItems => _popupStateKey.currentState?.getLoadedItems ?? [];
+  List<T> get popupGetItems =>
+      _popupStateKey.currentState?.getLoadedItems ?? [];
 
   void updatePopupState() => _popupStateKey.currentState?.setState(() {});
 }