+++ lib/src/properties/dropdown_props.dart
@@ -2,9 +2,10 @@ import 'package:dropdown_search/dropdown_search.dart';
 import 'package:dropdown_search/src/widgets/props/icon_button_props.dart';
 import 'package:flutter/material.dart';
 
-typedef DropdownButtonAnimationBuilder = Widget Function(Widget child, bool isOpen);
+typedef DropdownButtonAnimationBuilder = Widget Function(
+    Widget child, bool isOpen);
 
-Widget defaultAnimationBuilder(child, isOpen) {
+Widget defaultAnimationBuilder(Widget child, bool isOpen) {
   return AnimatedRotation(
     turns: isOpen ? .5 : 1,
     duration: Duration(milliseconds: 300),