+++ lib/src/widgets/props/chip_props.dart
@@ -129,7 +129,8 @@ class ChipProps {
       deleteIcon: deleteIcon ?? this.deleteIcon,
       onDeleted: onDeleted ?? this.onDeleted,
       deleteIconColor: deleteIconColor ?? this.deleteIconColor,
-      deleteButtonTooltipMessage: deleteButtonTooltipMessage ?? this.deleteButtonTooltipMessage,
+      deleteButtonTooltipMessage:
+          deleteButtonTooltipMessage ?? this.deleteButtonTooltipMessage,
       onPressed: onPressed ?? this.onPressed,
       pressElevation: pressElevation ?? this.pressElevation,
       disabledColor: disabledColor ?? this.disabledColor,
@@ -143,7 +144,8 @@ class ChipProps {
       backgroundColor: backgroundColor ?? this.backgroundColor,
       padding: padding ?? this.padding,
       visualDensity: visualDensity ?? this.visualDensity,
-      materialTapTargetSize: materialTapTargetSize ?? this.materialTapTargetSize,
+      materialTapTargetSize:
+          materialTapTargetSize ?? this.materialTapTargetSize,
       elevation: elevation ?? this.elevation,
       shadowColor: shadowColor ?? this.shadowColor,
       surfaceTintColor: surfaceTintColor ?? this.surfaceTintColor,
@@ -153,7 +155,8 @@ class ChipProps {
       avatarBorder: avatarBorder ?? this.avatarBorder,
       iconTheme: iconTheme ?? this.iconTheme,
       avatarBoxConstraints: avatarBoxConstraints ?? this.avatarBoxConstraints,
-      deleteIconBoxConstraints: deleteIconBoxConstraints ?? this.deleteIconBoxConstraints,
+      deleteIconBoxConstraints:
+          deleteIconBoxConstraints ?? this.deleteIconBoxConstraints,
       chipAnimationStyle: chipAnimationStyle ?? this.chipAnimationStyle,
       color: color ?? this.color,
       mouseCursor: mouseCursor ?? this.mouseCursor,
@@ -173,7 +176,8 @@ class ChipProps {
       deleteIcon: other.deleteIcon ?? deleteIcon,
       onDeleted: other.onDeleted ?? onDeleted,
       deleteIconColor: other.deleteIconColor ?? deleteIconColor,
-      deleteButtonTooltipMessage: other.deleteButtonTooltipMessage ?? deleteButtonTooltipMessage,
+      deleteButtonTooltipMessage:
+          other.deleteButtonTooltipMessage ?? deleteButtonTooltipMessage,
       onPressed: other.onPressed ?? onPressed,
       pressElevation: other.pressElevation ?? pressElevation,
       disabledColor: other.disabledColor ?? disabledColor,
@@ -187,7 +191,8 @@ class ChipProps {
       backgroundColor: other.backgroundColor ?? backgroundColor,
       padding: other.padding ?? padding,
       visualDensity: other.visualDensity ?? visualDensity,
-      materialTapTargetSize: other.materialTapTargetSize ?? materialTapTargetSize,
+      materialTapTargetSize:
+          other.materialTapTargetSize ?? materialTapTargetSize,
       elevation: other.elevation ?? elevation,
       shadowColor: other.shadowColor ?? shadowColor,
       surfaceTintColor: other.surfaceTintColor ?? surfaceTintColor,
@@ -197,7 +202,8 @@ class ChipProps {
       avatarBorder: other.avatarBorder,
       iconTheme: other.iconTheme ?? iconTheme,
       avatarBoxConstraints: other.avatarBoxConstraints ?? avatarBoxConstraints,
-      deleteIconBoxConstraints: other.deleteIconBoxConstraints ?? deleteIconBoxConstraints,
+      deleteIconBoxConstraints:
+          other.deleteIconBoxConstraints ?? deleteIconBoxConstraints,
       chipAnimationStyle: other.chipAnimationStyle ?? chipAnimationStyle,
       color: other.color ?? color,
       mouseCursor: other.mouseCursor ?? mouseCursor,